+++ module/item.js
@@ -1,11 +1,10 @@
-import {EntitySheetHelper} from "./helper.js";
+import { EntitySheetHelper } from "./helper.js";
 
 /**
  * Extend the base Item document to support attributes and groups with a custom template creation dialog.
  * @extends {Item}
  */
 export class SimpleItem extends Item {
-
   /** @inheritdoc */
   prepareDerivedData() {
     super.prepareDerivedData();
@@ -17,7 +16,7 @@ export class SimpleItem extends Item {
   /* -------------------------------------------- */
 
   /** @override */
-  static async createDialog(data={}, options={}) {
+  static async createDialog(data = {}, options = {}) {
     return EntitySheetHelper.createDialog.call(this, data, options);
   }
 
@@ -28,6 +27,6 @@ export class SimpleItem extends Item {
    * @type {boolean}
    */
   get isTemplate() {
-    return !!this.getFlag("worldbuilding", "isTemplate");
+    return !!this.getFlag("varlyn5e", "isTemplate");
   }
 }