+++ module/macro.js
@@ -1,21 +1,21 @@
 /**
  * Create a Macro from an attribute drop.
- * Get an existing worldbuilding macro if one exists, otherwise create a new one.
+ * Get an existing Varlyn5e macro if one exists, otherwise create a new one.
  * @param {Object} data     The dropped data
  * @param {number} slot     The hotbar slot to use
  * @returns {Promise}
  */
-export async function createWorldbuildingMacro(data, slot) {
-  if ( !data.roll || !data.label ) return false;
+export async function createVarlyn5eMacro(data, slot) {
+  if (!data.roll || !data.label) return false;
   const command = `const roll = new Roll("${data.roll}", actor ? actor.getRollData() : {});
   roll.toMessage({speaker, flavor: "${data.label}"});`;
-  let macro = game.macros.find(m => (m.name === data.label) && (m.command === command));
+  let macro = game.macros.find((m) => m.name === data.label && m.command === command);
   if (!macro) {
     macro = await Macro.create({
       name: data.label,
       type: "script",
       command: command,
-      flags: { "worldbuilding.attrMacro": true }
+      flags: { "varlyn5e.attrMacro": true },
     });
   }
   game.user.assignHotbarMacro(macro, slot);