+++ module/templates.js
@@ -7,8 +7,8 @@ export const preloadHandlebarsTemplates = async function () {
   // Define template paths to load
   const templatePaths = [
     // Attribute list partial.
-    "systems/worldbuilding/templates/parts/sheet-attributes.html",
-    "systems/worldbuilding/templates/parts/sheet-groups.html",
+    "systems/varlyn5e/templates/parts/sheet-attributes.html",
+    "systems/varlyn5e/templates/parts/sheet-groups.html",
   ];
 
   // Load the template parts